requirements.txt
langgraph
langchain
langchain-core
langchain-community
python-dotenv
requests
pydantic
pydantic-settings
httpx
fastapi
uvicorn
streamlit
sqlalchemy
requests>=2.31.0
python-multipart>=0.0.6
pdfplumber>=0.10.0
python-docx>=1.0.0
streamlit